+++ src/test/java/subway/integration/StationIntegrationTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import subway.dto.StationResponse;
+import subway.ui.dto.StationResponse;
 
 import java.util.HashMap;
 import java.util.List;
